chore: Update rss recipe to use handlers again, pull out and test RSS parsing
