+++ recipes/rss.tsx
@@ -1,138 +1,27 @@
 /// <cts-enable />
 import {
+  Cell,
   cell,
   Default,
   derive,
   fetchData,
-  lift,
+  handler,
   NAME,
   recipe,
   str,
   UI,
 } from "commontools";
-import { DOMParser, type Element } from "./dom-parser.ts";
+import { type FeedItem, parseRSSFeed } from "./rss-utils.ts";
 
 interface Settings {
   feedUrl: Default<string, "">;
   limit: Default<number, 100>;
 }
 
-type FeedItem = {
-  id: string;
-  title: string;
-  link: string;
-  description: string;
-  pubDate: string;
-  author: string;
-  content: string;
-};
-
-function parseRSSFeed(
-  textXML: string,
-  maxResults: number = 100,
-  existingIds: Set<string>,
-): FeedItem[] {
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(textXML, "text/xml");
-  // Helper function to get text content from an element
-  const getTextContent = (element: Element | null, tagName: string) => {
-    const el = element?.getElementsByTagName(tagName)[0];
-    return el?.textContent?.trim() || "";
-  };
-
-  // Helper function to get attribute value
-  const getAttributeValue = (
-    element: Element | null,
-    tagName: string,
-    attrName: string,
-  ) => {
-    const el = element?.getElementsByTagName(tagName)[0];
-    return el?.getAttribute(attrName) || "";
-  };
-
-  const retrievedItems: FeedItem[] = [];
-
-  // Check if it's an Atom feed
-  const isAtom = doc.getElementsByTagName("feed").length !== 0;
-
-  if (isAtom) {
-    // Parse Atom feed
-    const entries = doc.getElementsByTagName("entry");
-
-    for (let i = 0; i < Math.min(entries.length, maxResults); i++) {
-      const entry = entries[i];
-
-      // In Atom, id is mandatory
-      const id = getTextContent(entry, "id") || Math.random().toString(36);
-
-      // Skip if we already have this item
-      if (existingIds.has(id)) {
-        continue;
-      }
-
-      // Parse link - in Atom links are elements with href attributes
-      const link = getAttributeValue(entry, "link", "href");
-
-      // For content, check content tag first, then summary
-      const content = getTextContent(entry, "content") ||
-        getTextContent(entry, "summary");
-
-      // For author, it might be nested as <author><name>Author</name></author>
-      let author = "";
-      const authorEl = entry.getElementsByTagName("author")[0];
-      if (authorEl) {
-        author = getTextContent(authorEl, "name");
-      }
-
-      // For pubDate, Atom uses <published> or <updated>
-      const pubDate = getTextContent(entry, "published") ||
-        getTextContent(entry, "updated");
-
-      retrievedItems.push({
-        id,
-        title: getTextContent(entry, "title"),
-        link,
-        description: getTextContent(entry, "summary"),
-        pubDate,
-        author,
-        content,
-      });
-    }
-  } else {
-    // Parse RSS feed
-    const rssItems = doc.getElementsByTagName("item");
-
-    for (let i = 0; i < Math.min(rssItems.length, maxResults); i++) {
-      const item = rssItems[i];
-
-      const id = getTextContent(item, "guid") ||
-        getTextContent(item, "link") ||
-        Math.random().toString(36);
-
-      if (existingIds.has(id)) {
-        continue;
-      }
-
-      retrievedItems.push({
-        id,
-        title: getTextContent(item, "title"),
-        link: getTextContent(item, "link"),
-        description: getTextContent(item, "description"),
-        pubDate: getTextContent(item, "pubDate"),
-        author: getTextContent(item, "author"),
-        content: getTextContent(item, "content:encoded") ||
-          getTextContent(item, "description"),
-      });
-    }
-  }
-
-  return retrievedItems;
-}
-
-const feedUpdater = lift<{
-  items: FeedItem[];
+const feedUpdater = handler<never, {
+  items: Cell<FeedItem[]>;
   settings: Settings;
-}>(({ settings, items }) => {
+}>((_, { items, settings }) => {
   if (!settings.feedUrl) {
     console.warn("no feed URL provided");
     return;
@@ -140,13 +29,13 @@ const feedUpdater = lift<{
 
   const query = fetchData({ url: settings.feedUrl, mode: "text" });
   return derive(
-    { items, result: query.result, limit: settings.limit },
-    ({ result, limit, items }) => {
-      if (!result || typeof result !== "string") return;
+    { items, query, limit: settings.limit },
+    ({ query, limit, items }) => {
+      if (!query.result || typeof query.result !== "string") return;
       const newEntries = parseRSSFeed(
-        result as string,
+        query.result as string,
         limit,
-        new Set(items.map((item) => item.id)),
+        new Set(items.get().map((item) => item.id)),
       );
       items.push(...newEntries);
     },
